Clone framework and standard extras via adapt-cli devinstall
Uses nodegit to perform git clone on the framework and again on each
plugin in adapt.json. Looks up the url of the repository using Bower
(assumes that the package url is a git repo, which may not be globally
true). Doesn't do any fancy version checking, assumes that you're able to
resolve any issues there yourself.